requirements.txt
discord.py[voice] >= 2.2.0
discord-racket >= 0.0.13
yt-dlp